+++ pyfar/signals/__init__.py
@@ -1,8 +1,10 @@
 # -*- coding: utf-8 -*-
 
-from .signals import (
-    sine, impulse, noise, pulsed_noise,
-    linear_sweep, exponential_sweep)
+from .deterministic import (
+    sine, impulse, linear_sweep, exponential_sweep)
+
+from .stochastic import (
+    noise, pulsed_noise)
 
 __all__ = [
     'sine', 'impulse', 'noise', 'pulsed_noise',
